refactor: move header components of home page to header folder
